#135 Handling From Updates at the Reducer Level
